Added variant type for ores
