+++ src/main/java/motherlode/base/api/varianttypes/wood/StrippedBlockMap.java
@@ -1,0 +1,10 @@
+package motherlode.base.api.varianttypes.wood;
+
+import net.minecraft.block.Block;
+import motherlode.base.impl.StrippedBlockMapImpl;
+
+public interface StrippedBlockMap {
+    StrippedBlockMap INSTANCE = StrippedBlockMapImpl.INSTANCE;
+
+    void addStrippedBlock(Block block, Block strippedBlock);
+}
