--- src/main/java/motherlode/base/api/woodtype/StrippedBlockMap.java
@@ -1,10 +1,0 @@
-package motherlode.base.api.woodtype;
-
-import net.minecraft.block.Block;
-import motherlode.base.impl.StrippedBlockMapImpl;
-
-public interface StrippedBlockMap {
-    StrippedBlockMap INSTANCE = StrippedBlockMapImpl.INSTANCE;
-
-    void addStrippedBlock(Block block, Block strippedBlock);
-}
